**feature : Add Differents Window**
_Commit Description:_

Add posibility to switch window with Key (A Z E)

Add the Drafting Window

Improve PlayerRenderer to allow creation of Renderer specialized for Drafting Window.

Add resources for Drafting Window (pfp_draft.png and player_draft.png)

Add the enumeration Window which is used for define the window displayed, to be displayed.

[CR #038]